API

NonlinearSolvers.NonlinearSolvers — Module

NonlinearSolvers

A set of solvers for systems of non-linear equations

Numerical Methods

NonlinearSolvers.NewtonsMethod — Type

NewtonsMethod(
        f!::F!,
        j!::J!,
        x_init::A,
    ) where {F! <: Function, J! <: Function, A <: AbstractArray}

A non-linear system of equations type.

Fields

f!

Function to find the root of

j!

Jacobian of f!

x_init

Initial guess

x1

Storage

J

Storage

J⁻¹

Storage

F

Storage

NonlinearSolvers.NewtonsMethodAD — Type

NewtonsMethodAD(f!::F!, x_init::A) where {F!, A <: AbstractArray}

A non-linear system of equations type.

Fields

f!

Function to find the root of

x_init

Initial guess

x1

Storage

J

Storage

J⁻¹

Storage

F

Storage

Solve

NonlinearSolvers.solve! — Function

solve!(
    method::AbstractNonlinearSolverMethod{FT},
    soltype::SolutionType = CompactSolution(),
    tol::Union{Nothing, AbstractTolerance} = nothing,
    maxiters::Union{Nothing, Int} = 10_000,
)

Solve the non-linear system given

method the numerical method

soltype the solution type (CompactSolution or VerboseSolution)

tol the stopping tolerance

maxiters the maximum number of iterations to perform

Solution types

NonlinearSolvers.CompactSolution — Type

CompactSolution <: SolutionType

Used to return a CompactSolutionResults

NonlinearSolvers.VerboseSolution — Type

VerboseSolution <: SolutionType

Used to return a VerboseSolutionResults

Note

VerboseSolution is designed for debugging purposes only, and is not GPU-friendly.

Results types

NonlinearSolvers.CompactSolutionResults — Type

CompactSolutionResults{AT} <: AbstractSolutionResults{AT}

Result returned from find_zero when CompactSolution is passed as the soltype.

NonlinearSolvers.VerboseSolutionResults — Type

VerboseSolutionResults{AT} <: AbstractSolutionResults{AT}

Result returned from find_zero when VerboseSolution is passed as the soltype.

Stopping conditions (tolerances)

NonlinearSolvers.ResidualTolerance — Type

ResidualTolerance

A tolerance type based on the residual of the equation $f(x) = 0$

NonlinearSolvers.SolutionTolerance — Type

SolutionTolerance

A tolerance type based on the solution $x$ of the equation $f(x) = 0$